refactor: 각 타입별 카드 분리, update: reset.css 폰트 수정
