+++ client/src/styles/reset.css
@@ -1,4 +1,4 @@
-@import url('https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700;800;900;');
+@import url('https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@100;300;400;500;700;900&display=swap');
 
 html,
 body,
@@ -104,7 +104,7 @@ section {
 }
 body {
   line-height: 1;
-  font-family: 'Inter', sans-serif;
+  font-family: 'NotoSansKR', sans-serif;
 }
 a {
   color: black;
